update sample to use new logic components
