--- project/source/collision_benchmark/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(${GAME_EXE}
-    PRIVATE
-        ${ROOT_DIR}/source/collision_benchmark/CollisionBenchmark.cpp
-)
