--- project/source/collision_events/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(${GAME_EXE}
-    PRIVATE
-        ${ROOT_DIR}/source/collision_events/CollisionEvents.cpp
-)
