--- project/source/collision_events/PrefabSelector.h
@@ -1,86 +1,0 @@
-#pragma once
-
-#include "ecs/Registry.h"
-#include "ecs/component/Collider.h"
-#include "shared/Prefabs.h"
-#include "WasdController.h"
-#include "CollisionLogger.h"
-
-namespace nc::sample
-{
-    class PrefabSelector : public AutoComponent
-    {
-        public:
-            PrefabSelector(Entity entity, Registry* registry);
-
-            void FrameUpdate(float) override;
-            void Select(ColliderType type);
-
-        private:
-            Registry* m_registry;
-            Entity m_currentObject;
-            ColliderType m_typeToSpawn;
-            bool m_doSpawn;
-    };
-
-    PrefabSelector::PrefabSelector(Entity entity, Registry* registry)
-        : AutoComponent{entity},
-          m_registry{registry},
-          m_currentObject{Entity::Null()},
-          m_typeToSpawn{ColliderType::Box},
-          m_doSpawn{true}
-    {
-    }
-
-    void PrefabSelector::FrameUpdate(float)
-    {
-        if(!m_doSpawn)
-            return;
-        
-        m_doSpawn = false;
-
-        if(m_currentObject.Valid())
-        {
-            m_registry->Remove<Collider>(m_currentObject);
-            m_registry->Remove<Entity>(m_currentObject);
-        }
-
-        switch(m_typeToSpawn)
-        {
-            case ColliderType::Box:
-            {
-                m_currentObject = prefab::Create(m_registry, prefab::Resource::CubeBlue, {.tag = "Movable Object"});
-                m_registry->Add<Collider>(m_currentObject, BoxProperties{}, false);
-                break;
-            }
-            case ColliderType::Sphere:
-            {
-                m_currentObject = prefab::Create(m_registry, prefab::Resource::SphereBlue, {.tag = "Movable Object"});
-                m_registry->Add<Collider>(m_currentObject, SphereProperties{}, false);
-                break;
-            }
-            case ColliderType::Capsule:
-            {
-                m_currentObject = prefab::Create(m_registry, prefab::Resource::CapsuleBlue, {.tag = "Movable Object"});
-                m_registry->Add<Collider>(m_currentObject, CapsuleProperties{}, false);
-                break;
-            }
-            case ColliderType::Hull:
-            {
-                m_currentObject = prefab::Create(m_registry, prefab::Resource::DiscBlue, {.tag = "Movable Object"});
-                m_registry->Add<Collider>(m_currentObject, HullProperties{.assetPath = "coin.nca"}, false);
-                break;
-            }
-        }
-
-        m_registry->Add<PhysicsBody>(m_currentObject, PhysicsProperties{});
-        m_registry->Add<WasdController>(m_currentObject, m_registry, 2.0f);
-        m_registry->Add<CollisionLogger>(m_currentObject, m_registry);
-    }
-
-    void PrefabSelector::Select(ColliderType type)
-    {
-        m_typeToSpawn = type;
-        m_doSpawn = true;
-    }
-}
