--- project/source/jare_scratch/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(${GAME_EXE}
-    PRIVATE 
-        ${ROOT_DIR}/source/jare_scratch/VulkanScene.cpp
-)
