--- project/source/joints_test/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(${GAME_EXE}
-    PRIVATE
-        ${ROOT_DIR}/source/joints_test/JointsTest.cpp
-)
