--- project/source/rendering_benchmark/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(${GAME_EXE}
-    PRIVATE
-        ${ROOT_DIR}/source/rendering_benchmark/RenderingBenchmark.cpp
-)
