+++ project/source/scenes/CMakeLists.txt
@@ -1,0 +1,11 @@
+target_sources(${GAME_EXE}
+    PRIVATE
+        ${ROOT_DIR}/source/scenes/ClickEvents.cpp
+        ${ROOT_DIR}/source/scenes/CollisionEvents.cpp
+        ${ROOT_DIR}/source/scenes/CollisionBenchmark.cpp
+        ${ROOT_DIR}/source/scenes/JareTestScene.cpp
+        ${ROOT_DIR}/source/scenes/JointsTest.cpp
+        ${ROOT_DIR}/source/scenes/RenderingBenchmark.cpp
+        ${ROOT_DIR}/source/scenes/SpawnTest.cpp
+        ${ROOT_DIR}/source/scenes/Worms.cpp
+)
