+++ project/source/scenes/CollisionBenchmark.cpp
@@ -1,7 +1,7 @@
 #include "CollisionBenchmark.h"
 #include "NcEngine.h"
 #include "imgui/imgui.h"
-#include "shared/FPSTracker.h"
+#include "shared/Attachments.h"
 #include "shared/spawner/Spawner.h"
 
 namespace
@@ -67,7 +67,7 @@ namespace nc::sample
     void CollisionBenchmark::Load(NcEngine* engine)
     {
         auto* registry = engine->Registry();
-        
+
         // Setup
         m_sceneHelper.Setup(engine, false, true, Widget);
 
@@ -89,15 +89,18 @@ namespace nc::sample
 
         // Dynamic Cube Spawner
         auto dynamicSpawnerHandle = registry->Add<Entity>({.tag = "DynamicCubeSpawner"});
-        auto dynamicSpawner = registry->Add<Spawner>(dynamicSpawnerHandle, registry, prefab::Resource::CubeGreen, spawnBehavior, spawnExtension);
+        auto dynamicSpawner = registry->Add<Spawner>(dynamicSpawnerHandle, prefab::Resource::CubeGreen, spawnBehavior, spawnExtension);
+        registry->Add<FrameLogic>(dynamicSpawnerHandle, InvokeAttachment<Spawner>{});
 
         // Static Cube Spawner
         spawnBehavior.flags = Entity::Flags::Static;
         auto staticSpawnerHandle = registry->Add<Entity>({.tag = "StaticCubeSpawner"});
-        auto staticSpawner = registry->Add<Spawner>(staticSpawnerHandle, registry, prefab::Resource::CubeRed, spawnBehavior, spawnExtension);
+        auto staticSpawner = registry->Add<Spawner>(staticSpawnerHandle, prefab::Resource::CubeRed, spawnBehavior, spawnExtension);
+        registry->Add<FrameLogic>(staticSpawnerHandle, InvokeAttachment<Spawner>{});
 
         auto fpsTrackerHandle = registry->Add<Entity>({.tag = "FpsTracker"});
         auto fpsTracker = registry->Add<FPSTracker>(fpsTrackerHandle);
+        registry->Add<FrameLogic>(fpsTrackerHandle, InvokeAttachment<FPSTracker>{});
 
         // Lights
         auto lvHandle = registry->Add<Entity>({.position = Vector3{0.0f, 3.4f, 1.3f}, .tag = "Point Light 1"});
