+++ project/source/scenes/JointsTest.cpp
@@ -1,9 +1,10 @@
 #include "JointsTest.h"
 #include "NcEngine.h"
-#include "shared/Prefabs.h"
-#include "ForceBasedController.h"
-#include "shared/SceneNavigationCamera.h"
+#include "ecs/InvokeAttachment.h"
 #include "physics/PhysicsSystem.h"
+#include "shared/Attachments.h"
+#include "shared/GameLogic.h"
+#include "shared/Prefabs.h"
 
 namespace nc::sample
 {
@@ -17,6 +18,7 @@ namespace nc::sample
         // Camera
         auto cameraHandle = registry->Add<Entity>({.position = Vector3{0.0f, 6.1f, -6.5f}, .rotation = Quaternion::FromEulerAngles(0.7f, 0.0f, 0.0f), .tag = "Main Camera"});
         auto camera = registry->Add<SceneNavigationCamera>(cameraHandle, 0.05f, 0.005f, 1.4f);
+        registry->Add<FrameLogic>(cameraHandle, InvokeAttachment<SceneNavigationCamera>{});
         engine->MainCamera()->Set(camera);
 
         // Lights
@@ -31,7 +33,7 @@ namespace nc::sample
             auto head = prefab::Create(registry, prefab::Resource::CubeGreen, {.tag = "Movable Object"});
             registry->Add<Collider>(head, BoxProperties{}, false);
             registry->Add<PhysicsBody>(head, PhysicsProperties{.mass = 5.0f});
-            registry->Add<ForceBasedController>(head, registry);
+            registry->Add<FixedLogic>(head, ForceBasedMovement);
 
             auto segment1 = prefab::Create(registry, prefab::Resource::CubeGreen, {.position = Vector3{0.0f, 0.0f, -0.9f}, .scale = Vector3::Splat(0.8f)});
             registry->Add<Collider>(segment1, BoxProperties{}, false);
