+++ project/source/shared/CMakeLists.txt
@@ -1,7 +1,8 @@
 target_sources(${GAME_EXE}
     PRIVATE
-        ${ROOT_DIR}/source/shared/EdgePanCamera.cpp
+        ${ROOT_DIR}/source/shared/Attachments.cpp
         ${ROOT_DIR}/source/shared/GameLog.cpp
+        ${ROOT_DIR}/source/shared/GameLogic.cpp
         ${ROOT_DIR}/source/shared/Prefabs.cpp
         ${ROOT_DIR}/source/shared/SampleUI.cpp
 )
