--- project/source/shared/EdgePanCamera.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "ecs/component/Camera.h"
-
-namespace nc::sample
-{
-    class EdgePanCamera : public Camera
-    {
-        public:
-            EdgePanCamera(Entity entity);
-            void FrameUpdate(float dt) override;
-
-        private:
-            float m_lastFrameZoom;
-            Vector3 m_lastFrameTranslation = Vector3::Zero();
-    };
-}
