+++ project/source/shared/Spawner/Spawner.h
@@ -1,8 +1,7 @@
 #pragma once
 #include "SpawnPropertyGenerator.h"
 #include "shared/Prefabs.h"
-#include "shared/ConstantTranslation.h"
-#include "shared/ConstantRotation.h"
+#include "shared/Attachments.h"
 
 #include <functional>
 
@@ -10,27 +9,25 @@ namespace nc::sample
 {
     /** A prefab spawner configurable with a SpawnBehavior. If provided, extension
         will be applied to each handle after creation. */
-    class Spawner : public AutoComponent
+    class Spawner : public StateAttachment
     {
         public:
             using SpawnExtension = std::function<void(Entity)>;
 
             Spawner(Entity entity,
-                    Registry* registry,
                     prefab::Resource resource,
                     SpawnBehavior behavior,
                     SpawnExtension extension = nullptr);
-            void FrameUpdate(float) override;
+            void Run(Entity, Registry*, float);
             void StageSpawn(unsigned count = 1u);
-            void Spawn(unsigned count = 1u);
+            void Spawn(Registry* registry, unsigned count = 1u);
             void StageDestroy(unsigned count = 1u);
-            void Destroy(unsigned count = 1u);
+            void Destroy(Registry* registry, unsigned count = 1u);
             void SetPrefab(prefab::Resource resource);
             const std::vector<Entity>& GetHandles() const;
             int GetObjectCount() const;
         
         private:
-            Registry* m_registry;
             SpawnExtension m_extension;
             std::vector<Entity> m_entities;
             SpawnPropertyGenerator m_generator;
@@ -44,12 +41,10 @@ namespace nc::sample
     };
 
     inline Spawner::Spawner(Entity entity,
-                            Registry* registry,
                             prefab::Resource resource,
                             SpawnBehavior behavior,
                             SpawnExtension extension)
-        : AutoComponent{entity},
-          m_registry{registry},
+        : StateAttachment{entity},
           m_extension{extension},
           m_entities{},
           m_generator{behavior},
@@ -64,20 +59,20 @@ namespace nc::sample
     {
     }
 
-    inline void Spawner::FrameUpdate(float)
+    inline void Spawner::Run(Entity, Registry* registry, float)
     {
         // Additions/Deletions are delayed until FrameUpdate because Spawn/Destroy are
         // callbacks from ui events, and modifying state in the middle of rendering can
         // cause problems.
         if(m_stagedAdditions)
         {
-            Spawn(m_stagedAdditions);
+            Spawn(registry, m_stagedAdditions);
             m_stagedAdditions = 0u;
         }
 
         if(m_stagedDeletions)
         {
-            Destroy(m_stagedDeletions);
+            Destroy(registry, m_stagedDeletions);
             m_stagedDeletions = 0u;
         }
     }
@@ -87,11 +82,11 @@ namespace nc::sample
         m_stagedAdditions = count;
     }
 
-    inline void Spawner::Spawn(unsigned count)
+    inline void Spawner::Spawn(Registry* registry, unsigned count)
     {
-        std::generate_n(std::back_inserter(m_entities), count, [this]()
+        std::generate_n(std::back_inserter(m_entities), count, [this, registry]()
         {
-            auto handle = prefab::Create(m_registry, m_resource,
+            auto handle = prefab::Create(registry, m_resource,
             {
                 .position = m_generator.Position(),
                 .rotation = Quaternion::FromEulerAngles(m_generator.Rotation()),
@@ -100,9 +95,18 @@ namespace nc::sample
             });
 
             if(m_applyConstantVelocity)
-                m_registry->Add<ConstantTranslation>(handle, m_registry, m_generator.Velocity());
+                registry->Add<ConstantTranslation>(handle, m_generator.Velocity());
             if(m_applyConstantRotation)
-                m_registry->Add<ConstantRotation>(handle, m_registry, m_generator.RotationAxis(), m_generator.Theta());
+                registry->Add<ConstantRotation>(handle, m_generator.RotationAxis(), m_generator.Theta());
+
+            registry->Add<FrameLogic>(handle, [](Entity self, Registry* registry, float dt)
+            {
+                if(auto* translation = registry->Get<ConstantTranslation>(self))
+                    translation->Run(self, registry, dt);
+                if(auto* rotation = registry->Get<ConstantRotation>(self))
+                    rotation->Run(self, registry, dt);
+            });
+
             if(m_extension)
                 m_extension(handle);
             return handle;
@@ -114,11 +118,11 @@ namespace nc::sample
         m_stagedDeletions = count;
     }
 
-    inline void Spawner::Destroy(unsigned count)
+    inline void Spawner::Destroy(Registry* registry, unsigned count)
     {
         while(!m_entities.empty() && count--)
         {
-            m_registry->Remove<Entity>(m_entities.back());
+            registry->Remove<Entity>(m_entities.back());
             m_entities.pop_back();
         }
     }
