--- project/source/solar_system/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(${GAME_EXE}
-    PRIVATE
-        # ${ROOT_DIR}/source/solar_system/SolarSystem.cpp
-)
