--- project/source/spawn_test/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(${GAME_EXE}
-    PRIVATE
-        ${ROOT_DIR}/source/spawn_test/SpawnTest.cpp
-)
