--- project/source/worms/CMakeLists.txt
@@ -1,5 +1,0 @@
-target_sources(${GAME_EXE}
-    PRIVATE
-        ${ROOT_DIR}/source/worms/MouseFollower.cpp
-        ${ROOT_DIR}/source/worms/Worms.cpp
-)
